feat:#32 트러블슈팅게시판레이아웃
