+++ src/assets/components/StudyCategory/DetailCategory.jsx
