+++ src/assets/components/TroubleShooting/TitleSection.jsx
@@ -1,17 +1,15 @@
 import React from "react";
 import styled from "styled-components";
-import Logo from "../../assets/logos/logo.svg";
-import BackgroundImage from "../../assets/images/titleBackground.png";
+import BackgroundImage from "../../../assets/images/titleBackground.png";
 
 const TitleSection = () => {
   return (
     <TitleContainer>
-      <LogoImg src={Logo} alt="Study Room Logo" />
       <Title>스터디룸</Title>
       <NavButtons>
         <NavButton>스터디 홈</NavButton>
-        <NavButton>트러블 슈팅 게시판</NavButton>
-        <NavButton>정보나눔 게시판</NavButton>
+        <NavButton>트러블슈팅</NavButton>
+        <NavButton>정보나눔</NavButton>
         <NavButton>채팅방</NavButton>
       </NavButtons>
     </TitleContainer>
@@ -22,7 +20,6 @@ export default TitleSection;
 
 const TitleContainer = styled.div`
   text-align: center;
-  margin: 40px 0;
   background-image: url(${BackgroundImage});
   background-size: cover;
   background-position: center;
@@ -30,35 +27,34 @@ const TitleContainer = styled.div`
   border-radius: 10px;
 `;
 
-const LogoImg = styled.img`
-  width: 40px;
-  height: 40px;
-`;
-
 const Title = styled.h1`
   font-size: 20px;
   color: #8e59ff;
-  margin: 10px 0;
+  margin-bottom: 30px;
+  font-weight: 800;
 `;
 
 const NavButtons = styled.div`
   display: flex;
   justify-content: center;
-  margin-top: 40px;
   gap: 20px;
 `;
 
 const NavButton = styled.button`
-  background-color: #8e59ff;
   background-color: rgba(142, 89, 255, 0.6);
   color: white;
   border: none;
   padding: 10px 20px;
-  border-radius: 5px;
-  width: 224px;
+  border-radius: 14px;
   cursor: pointer;
+  width: 140px;
+  text-align: center;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  transition: background-color 0.3s ease;
 
   &:hover {
-    background-color: #8e59ff;
+    background-color: rgba(142, 89, 255, 1);
   }
 `;
