+++ src/store/someReducer.jsx
@@ -1,0 +1,13 @@
+const initialState = {
+  // 초기 상태 정의
+};
+
+const someReducer = (state = initialState, action) => {
+  switch (action.type) {
+    // 액션 타입에 따른 상태 변경 로직
+    default:
+      return state;
+  }
+};
+
+export default someReducer; // default export 추가
