+++ src/store/store.jsx
@@ -1,7 +1,6 @@
-import { configureStore } from '@reduxjs/toolkit';
+import { configureStore } from "@reduxjs/toolkit";
+import rootReducer from "./rootReducer";
 
 export const store = configureStore({
-    reducer: {
-        
-    },
+  reducer: rootReducer,
 });
